third_party/WebKit/LayoutTests/bluetooth/getCharacteristic.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="bluetooth-helpers.js"></script>
<script>
test(function(t) { assert_true(window.testRunner instanceof Object); t.done(); },
     "window.testRunner is required for the following tests.");

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
      return gattServer.getPrimaryService('generic_access');
  }).then(function(service) {
    testRunner.setBluetoothMockDataSet('EmptyAdapter');
    return service.getCharacteristic(generic_access.device_name.uuid).then(function() {
      assert_unreached("Device went out of range, should fail.");
    }, function(e) {
      assert_equals(e.name, "NetworkError");
      assert_equals(e.message, 'Bluetooth Device is no longer in range.');
    });
  });
}, 'Device goes out of range. Reject with NetworkError.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['generic_access']}]})
    .then(device => device.connectGATT())
    .then(gattService => gattService.getPrimaryService('generic_access'))
    .then(service => {
      testRunner.setBluetoothMockDataSet('MissingServiceGenericAccessAdapter');
      return service.getCharacteristic('device_name').then(() => {
        assert_unreached('Service was removed, should fail.');
      }, (error) => {
        assert_equals(error.name, 'InvalidStateError');
        assert_equals(error.message, 'GATT Service no longer exists.');
      });
    });
}, 'Services is removed. Reject with InvalidStateError.');


sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService('generic_access');
  }).then(function(service) {
    return Promise.all(
      [service.getCharacteristic(battery_service.battery_level.alias),
       service.getCharacteristic(battery_service.battery_level.name),
       service.getCharacteristic(battery_service.battery_level.uuid)]);
  }).then(function(characteristics) {
    characteristics.forEach(function(characteristic) {
      assert_equals(characteristic, null,
                    "Non existent characteristic should return null.");
    });
  });
}, 'Request for wrong characteristic. Should return null.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService('generic_access');
  }).then(function(service) {
    return Promise.all([service.getCharacteristic(generic_access.device_name.alias),
                        service.getCharacteristic(generic_access.device_name.name),
                        service.getCharacteristic(generic_access.device_name.uuid)]);
  }).then(function(characteristics){
    characteristics.forEach(function(characteristic) {
      assert_equals(characteristic.uuid, generic_access.device_name.uuid,
                    "Characteristic UUID should be the same as requested UUID.");
    });
  });
}, 'Request for characteristic. Should return right characteristic');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService('generic_access');
  }).then(function(services) {
    return Promise.all([services.getCharacteristic(generic_access.device_name.alias),
                        services.getCharacteristic(generic_access.device_name.alias),
                        services.getCharacteristic(generic_access.device_name.name),
                        services.getCharacteristic(generic_access.device_name.name),
                        services.getCharacteristic(generic_access.device_name.uuid),
                        services.getCharacteristic(generic_access.device_name.uuid)]);
  }).then(function(characteristics) {
    // TODO(ortuno): getCharacteristic should return the same object
    // if it was created earlier.
    // https://crbug.com/495270
    for (var i = 1; i < characteristics.length; i++) {
      assert_not_equals(characteristics[0], characteristics[i],
                        "Should return the same characteristic as the first call.");
    }
  });
}, 'Calls to get the same characteristic should return the same object.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
      return gattServer.getPrimaryService('generic_access');
  }).then(function(service) {
    return service.getCharacteristic("wrong_name").then(function() {
      assert_unreached("Wrong Characteristic name, should fail.");
    }, function(e) {
      assert_equals(e.name, "SyntaxError");
      assert_equals(e.message, "Failed to execute 'getCharacteristic' on 'BluetoothGATTService': Invalid Characteristic name: 'wrong_name'.");
    });
  });
}, 'Wrong Characteristic name. Reject with SyntaxError.');
</script>
